requirements.txt
flask==3.0.1
requests==2.32.3
jsonify==0.5
python-dotenv==1.1.0